+++ R/analysis-totals.R
@@ -54,7 +54,7 @@
 #' `meta(result)$variable`. Use `meta(result)` for additional metadata.
 #'
 #' @examples
-#' d <- as_survey_rep(acs_pums_wy, weights = pwgtp,
+#' d <- as_survey_repweights(acs_pums_wy, weights = pwgtp,
 #'                    repweights = pwgtp1:pwgtp80,
 #'                    type = "successive-difference")
 #'